# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/webhook": {
      "post": {
        "operationId": "WebhookController_handleStripeWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/stripe/product/{id}": {
      "delete": {
        "operationId": "StripeController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Stripe Operations"
        ]
      }
    },
...
    "/api/products": {
      "post": {
        "operationId": "ProductController_create",
        "summary": "Create a new product",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProductController_findAll",
        "summary": "Get all products",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ADDON",
                "PRODUCT",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/products/addons": {
      "get": {
        "operationId": "ProductController_findAddons",
        "summary": "Get all addon products",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Addon products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/products/products": {
      "get": {
        "operationId": "ProductController_findProducts",
        "summary": "Get all standalone products",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Standalone products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductController_update",
        "summary": "Update product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductController_remove",
        "summary": "Soft delete product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/products/{id}/features/{featureId}": {
      "post": {
        "operationId": "ProductController_addFeature",
        "summary": "Add feature to product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featureId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Feature added to product successfully"
          },
          "404": {
            "description": "Product or feature not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductController_removeFeature",
        "summary": "Remove feature from product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featureId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature removed from product successfully"
          },
          "404": {
            "description": "Product feature relationship not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/prices/product/{productId}": {
      "get": {
        "operationId": "PriceController_findByProduct",
        "summary": "Get prices by product ID",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product prices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "prices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Virtual Mail Hub",
    "description": "The Virtual Mail Hub API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Extra Recipient"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Additional recipient for mail forwarding"
          },
          "stripeProductId": {
            "type": "string",
            "description": "Stripe product ID",
            "example": "prod_1234567890"
          },
          "type": {
            "type": "string",
            "description": "Product type",
            "enum": [
              "ADDON",
              "PRODUCT",
              "OTHER"
            ],
            "example": "ADDON"
          },
          "imageUrl": {
            "type": "string",
            "description": "Product image URL",
            "example": "https://example.com/product-image.png"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is product active",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Product ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Extra Recipient"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Additional recipient for mail forwarding"
          },
          "stripeProductId": {
            "type": "string",
            "description": "Stripe product ID",
            "example": "prod_1234567890"
          },
          "type": {
            "type": "string",
            "description": "Product type",
            "enum": [
              "ADDON",
              "PRODUCT",
              "OTHER"
            ],
            "example": "ADDON"
          },
          "imageUrl": {
            "type": "string",
            "description": "Product image URL",
            "example": "https://example.com/product-image.png"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is product active",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is product deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "prices": {
            "description": "Product prices",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "productFeature": {
            "description": "Product features",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "planAddon": {
            "description": "Plan addons using this product",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "isActive",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Extra Recipient"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Additional recipient for mail forwarding"
          },
          "stripeProductId": {
            "type": "string",
            "description": "Stripe product ID",
            "example": "prod_1234567890"
          },
          "type": {
            "type": "string",
            "description": "Product type",
            "enum": [
              "ADDON",
              "PRODUCT",
              "OTHER"
            ],
            "example": "ADDON"
          },
          "imageUrl": {
            "type": "string",
            "description": "Product image URL",
            "example": "https://example.com/product-image.png"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is product active",
            "example": true,
            "default": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is product deleted",
            "example": false
          }
        }
      },
...
      "PriceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Price ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Price name",
            "example": "Monthly Extra Recipient"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Is default price",
            "example": false
          },
          "additionalFees": {
            "type": "number",
            "description": "Additional fees in cents",
            "example": 50
          },
          "stripePriceId": {
            "type": "string",
            "description": "Stripe price ID",
            "example": "price_1234567890"
          },
          "unit_amount": {
            "type": "number",
            "description": "Unit amount in cents",
            "example": 299
          },
          "currency": {
            "type": "string",
            "description": "Currency",
            "example": "USD"
          },
          "productId": {
            "type": "string",
            "description": "Product ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "description": {
            "type": "string",
            "description": "Price description",
            "example": "Monthly subscription for extra recipient"
          },
          "recurringId": {
            "type": "string",
            "description": "Recurring ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "active": {
            "type": "boolean",
            "description": "Is price active",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is price deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "product": {
            "type": "object",
            "description": "Product details"
          },
          "recurring": {
            "type": "object",
            "description": "Recurring details"
          },
          "planAddons": {
            "description": "Plan addons using this price",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "isDefault",
          "unit_amount",
          "currency",
          "productId",
          "active",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}